+++ src/main/java/com/alkemy/ong/controllers/CommentController.java
@@ -1,9 +1,8 @@
 package com.alkemy.ong.controllers;
 
-import com.alkemy.ong.dtos.requests.CommentPostRequestDTO;
-import com.alkemy.ong.dtos.requests.CommentPutRequestDTO;
+import com.alkemy.ong.dtos.requests.createAndUpdate.CommentPostRequestDTO;
+import com.alkemy.ong.dtos.requests.createAndUpdate.CommentPutRequestDTO;
 import com.alkemy.ong.dtos.responses.CommentDTO;
-import com.alkemy.ong.dtos.responses.CommentListDTO;
 import com.alkemy.ong.entities.Comment;
 import com.alkemy.ong.services.CommentService;
 import com.alkemy.ong.util.JwtUtil;
