+++ src/main/java/com/alkemy/ong/controllers/NewsController.java
@@ -1,10 +1,9 @@
 package com.alkemy.ong.controllers;
 
-import com.alkemy.ong.dtos.requests.NewPostPutRequestDTO;
+import com.alkemy.ong.dtos.requests.createAndUpdate.NewPostPutRequestDTO;
 import com.alkemy.ong.dtos.responses.NewDTO;
 import com.alkemy.ong.dtos.responses.NewsPageResponseDTO;
 import com.alkemy.ong.entities.News;
-import com.alkemy.ong.exceptions.BadRequestException;
 import com.alkemy.ong.exceptions.InvalidParameterException;
 import com.alkemy.ong.exceptions.NotFoundException;
 import com.alkemy.ong.services.INewsService;
