+++ src/main/java/com/alkemy/ong/controllers/OrganizationController.java
@@ -1,20 +1,16 @@
 package com.alkemy.ong.controllers;
 
-import com.alkemy.ong.dtos.requests.SlideRequest;
+import com.alkemy.ong.dtos.requests.createAndUpdate.SlideRequest;
 import com.alkemy.ong.entities.OrganizationEntity;
-import com.alkemy.ong.exceptions.ParamNotFound;
 import com.alkemy.ong.mapper.SlideMapper;
-import com.alkemy.ong.pojos.input.CreateOrganizationDTO;
-import com.alkemy.ong.pojos.input.UpdateOrganizationDTO;
+import com.alkemy.ong.dtos.requests.createAndUpdate.CreateOrganizationDTO;
+import com.alkemy.ong.dtos.requests.createAndUpdate.UpdateOrganizationDTO;
 import com.alkemy.ong.pojos.output.FindOrganizationDTO;
 import com.alkemy.ong.pojos.output.ListOrganizationDTO;
 import com.alkemy.ong.services.OrganizationService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.validation.ObjectError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
