--- src/main/java/com/alkemy/ong/dtos/requests/CommentPutRequestDTO.java
@@ -1,10 +1,0 @@
-package com.alkemy.ong.dtos.requests;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CommentPutRequestDTO {
-    private String body;
-}
