+++ src/main/java/com/alkemy/ong/dtos/requests/createAndUpdate/CommentPostRequestDTO.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.dtos.requests;
+package com.alkemy.ong.dtos.requests.createAndUpdate;
 
 import lombok.Builder;
 import lombok.Data;
@@ -9,8 +9,10 @@ import javax.validation.constraints.NotNull;
 
 @Data
 @Builder
+/**
+ * Se usa en create
+ */
 public class CommentPostRequestDTO {
-
     @NotNull
     @Min(value = 1, message = "News id cannot be less than one.")
     private Long news_id;
