+++ src/main/java/com/alkemy/ong/dtos/requests/createAndUpdate/ContactPostDTO.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.dtos.requests;
+package com.alkemy.ong.dtos.requests.createAndUpdate;
 
 import com.alkemy.ong.entities.Contact;
 import lombok.AllArgsConstructor;
@@ -7,29 +7,33 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Size;
 
+/**
+ * se usa en create, no tiene update
+ */
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 @Data
 public class ContactPostDTO {
 
-    @NotEmpty
+    @NotBlank
     @Size(min = 3, message = "Name should be at least three characters")
     private String name;
 
-    @NotEmpty
+    @NotBlank
     @Size(min = 10, message = "Phone number should be at least nine characters")
     private String phone;
 
-    @NotEmpty
+    @NotBlank
     @Email(message = "Email is not valid", regexp = "^(?=.{1,64}@)[A-Za-z0-9_-]+(\\.[A-Za-z0-9_-]+)*@"
             + "[^-][A-Za-z0-9-]+(\\.[A-Za-z0-9-]+)*(\\.[A-Za-z]{2,})$")
     private String email;
 
-    @NotEmpty
+    @NotBlank
     private String message;
 
     public Contact toContact(){
