+++ src/main/java/com/alkemy/ong/dtos/requests/createAndUpdate/MemberRequest.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.dtos.requests;
+package com.alkemy.ong.dtos.requests.createAndUpdate;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -7,6 +7,9 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 
+/**
+ * Se usa en create y put
+ */
 @Getter
 @Setter
 public class MemberRequest extends Socials{
