+++ src/main/java/com/alkemy/ong/dtos/requests/createAndUpdate/NewPostPutRequestDTO.java
@@ -1,24 +1,28 @@
-package com.alkemy.ong.dtos.requests;
+package com.alkemy.ong.dtos.requests.createAndUpdate;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
+/**
+ * Se usa en create y put
+ */
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class NewPostPutRequestDTO {
-    @NotNull(message = "Name may not be empty")
+    @NotBlank(message = "Name may not be empty")
     private String name;
 
-    @NotNull(message = "Content may not be empty")
+    @NotBlank(message = "Content may not be empty")
     private String content;
 
-    @NotNull(message = "Image may not be empty")
+    @NotBlank(message = "Image may not be empty")
     private String image;
 
     @NotNull(message = "Category may not be empty")
