+++ src/main/java/com/alkemy/ong/dtos/requests/createAndUpdate/SlideRequest.java
@@ -1,10 +1,13 @@
-package com.alkemy.ong.dtos.requests;
+package com.alkemy.ong.dtos.requests.createAndUpdate;
 
 import lombok.Data;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 
+/**
+ * put y create
+ */
 @Data
 public class SlideRequest {
     @NotBlank
