+++ src/main/java/com/alkemy/ong/services/INewsService.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.services;
 
-import com.alkemy.ong.dtos.requests.NewPostPutRequestDTO;
+import com.alkemy.ong.dtos.requests.createAndUpdate.NewPostPutRequestDTO;
 import com.alkemy.ong.dtos.responses.NewDTO;
 import com.alkemy.ong.entities.News;
 import com.alkemy.ong.exceptions.NotFoundException;
