+++ src/main/java/com/alkemy/ong/services/TestimonialService.java
@@ -1,7 +1,7 @@
 package com.alkemy.ong.services;
 
 import com.alkemy.ong.entities.TestimonialEntity;
-import com.alkemy.ong.pojos.input.TestimonialDTO;
+import com.alkemy.ong.dtos.requests.createAndUpdate.TestimonialDTO;
 import org.springframework.data.domain.Slice;
 
 public interface TestimonialService {
