+++ src/test/java/com/alkemy/ong/controllers/CategoryControllerTest.java
@@ -1,7 +1,7 @@
 package com.alkemy.ong.controllers;
 
 import com.alkemy.ong.dtos.requests.CategoryListRequestDTO;
-import com.alkemy.ong.dtos.requests.CategoryPostPutRequestDTO;
+import com.alkemy.ong.dtos.requests.createAndUpdate.CategoryPostPutRequestDTO;
 import com.alkemy.ong.dtos.responses.CategoryDTO;
 import com.alkemy.ong.entities.Category;
 import com.alkemy.ong.exceptions.BadRequestException;
@@ -9,18 +9,13 @@ import com.alkemy.ong.exceptions.ParamNotFound;
 import com.alkemy.ong.pojos.output.PageDTO;
 import com.alkemy.ong.repositories.CategoryRepository;
 import com.alkemy.ong.services.CategoryService;
-import com.alkemy.ong.services.impl.CategoryServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Slice;
-import org.springframework.data.domain.SliceImpl;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
