+++ src/test/java/com/alkemy/ong/controllers/ContactControllerTest.java
@@ -1,13 +1,12 @@
 package com.alkemy.ong.controllers;
 
-import com.alkemy.ong.dtos.requests.ContactPostDTO;
+import com.alkemy.ong.dtos.requests.createAndUpdate.ContactPostDTO;
 import com.alkemy.ong.dtos.responses.ContactListDTO;
 import com.alkemy.ong.entities.Contact;
 import com.alkemy.ong.services.ContactService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
